requirements.txt
boto3==1.35.5
botocore==1.35.5
pyyaml==6.0.2
python-dotenv==1.0.1
